amfit_ecommerce/requirements-dev.txt
-r requirements.txt

# Development Tools
pytest==7.4.3
pytest-django==4.7.0
black==23.12.0
flake8==6.1.0
isort==5.13.2
django-debug-toolbar==4.2.0
